[renderer] Remove useTopicManager and integrate topic subscription logic into MinutesAction
