+++ src/renderer/views/action/MinutesAction.ts
@@ -120,6 +120,7 @@ const prepareStoresTo = async (startTimestamp: number) => {
   ]);
 
   // == Subscribe stores ==
+  subscribeTopicInvokeQueue(startTimestamp);
   subscribeAssistantInvokeQueue(startTimestamp);
 
   // == Subscribe to reactflow ==
@@ -154,126 +155,25 @@ const subscribeAssistantInvokeQueue = (startTimestamp: number) => {
     }
   );
 };
-/*
-export function useTopicManager(): void {
-  const startTimestamp = useVBStore((state) => state.startTimestamp);
-  const topicAIConf = useMinutesStore(startTimestamp)(
-    (state) => state.topicAIConf
-  );
-  const topicState = useTopicStore((state) => state);
-  const getAgenda = useMinutesAgendaStore(startTimestamp)(
-    (state) => state.getAgenda
-  );
 
-  // process stocked prompts
-  useEffect(() => {
-    const targetIndex = topicState.prompts.findIndex((v) => !v.isRequested);
-    if (!topicState.processing && targetIndex > -1) {
-      const handleRequest = async () => {
-        topicState.startProcess();
-
-        // make a request
-        const currentTopic = topicState.prompts[targetIndex];
-        let lastResult = "";
-        if (topicState.res && topicState.res.data.topics.length > 0) {
-          const lastTopic =
-            topicState.res.data.topics[topicState.res.data.topics.length - 1];
-          lastResult = `${lastTopic.title}\n\n ${
-            Array.isArray(lastTopic.topic)
-              ? lastTopic.topic.join("\n")
-              : lastTopic.topic
-          }`;
-
-          if (lastTopic.seedData && lastTopic.seedData.agendaIdList) {
-            lastTopic.seedData.agendaIdList.forEach((agendaId) => {
-              const relatedAgenda = getAgenda(agendaId);
-              console.log(
-                "useTopicManager: lastTopic agendaIdList",
-                relatedAgenda
-              );
-            });
-          }
-        }
-
-        // agenda
-        if (currentTopic.seedData.agendaIdList) {
-          currentTopic.seedData.agendaIdList.forEach((agendaId) => {
-            const relatedAgenda = getAgenda(agendaId);
-            console.log(
-              "useTopicManager: currentTopic agendaIdList",
-              relatedAgenda
-            );
-          });
-        }
-
-        const prompt = `"""Immediate Previous Content\n${lastResult}\n"""\n\n"""Content\n${currentTopic.text}\n"""\n`;
-        const updatePrompts = [...topicState.prompts];
-        updatePrompts[targetIndex].isRequested = true;
-        await window.electron
-          .invoke(IPCInvokeKeys.GET_TOPIC, {
-            systemPrompt: EnglishTopicPrompt, //,vbState.topicAIConf.systemPrompt,
-            structuredOutputSchema: topicAIConf.structuredOutputSchema,
-            inputPrompt: prompt,
-            modelType: topicAIConf.modelType,
-            temperature: topicAIConf.temperature,
-            difyConf: topicAIConf.difyConf,
-            flowiseConf: topicAIConf.flowiseConf,
-          } as TopicInvokeParam)
-          .then((resJSON) => {
-            console.log("useTopicManager: getTopic", resJSON);
-            if (isLLMAnalyzedTopics(resJSON)) {
-              resJSON.topics.forEach((topic) => {
-                topic.id = uuidv4();
-                topic.seedData = currentTopic.seedData;
-                topic.agendaIds = currentTopic.seedData.agendaIdList ?? [];
-              });
-              const responsePrompt = topicState.prompts[targetIndex];
-              topicState.topicSeeds.forEach((seed) => {
-                if (seed == responsePrompt.seedData) {
-                  seed.requireUpdate = false;
-                }
-              });
-              topicState.resSuccess({
-                res: {
-                  data: resJSON,
-                },
-                prompts: updatePrompts,
-                topicSeed: topicState.topicSeeds,
-              });
-            } else {
-              console.error("resError: Invalid JSON data", resJSON);
-              const responsePrompt = topicState.prompts[targetIndex];
-              topicState.topicSeeds.forEach((seed) => {
-                if (seed == responsePrompt.seedData) {
-                  seed.requireUpdate = false;
-                }
-              });
-              topicState.resError({
-                error: new Error("Invalid JSON data"),
-                prompts: updatePrompts,
-                topicSeed: topicState.topicSeeds,
-              });
-            }
-          })
-          .catch((e) => {
-            console.error("resError :general error", e);
-            const responsePrompt = topicState.prompts[targetIndex];
-            topicState.topicSeeds.forEach((seed) => {
-              if (seed == responsePrompt.seedData) {
-                seed.requireUpdate = false;
-              }
-            });
-            topicState.resError({
-              error: e,
-              prompts: updatePrompts,
-              topicSeed: topicState.topicSeeds,
-            });
-          });
-      };
-
-      handleRequest();
+let unsubscribeTopicInvokeQueue: (() => void) | null = null;
+const subscribeTopicInvokeQueue = (startTimestamp: number) => {
+  const topicStore = useMinutesStore(startTimestamp);
+  if (unsubscribeTopicInvokeQueue) unsubscribeTopicInvokeQueue();
+  unsubscribeTopicInvokeQueue = topicStore.subscribe(
+    (state) => ({
+      topicProcessing: state.topicProcessing,
+      topicPrompts: state.topicPrompts,
+      topicRes: state.topicRes,
+    }),
+    (state) => {
+      topicStore.getState().processTopic();
+    },
+    {
+      equalityFn: (prev, next) =>
+        prev.topicProcessing === next.topicProcessing && // no change in processing
+        prev.topicPrompts === next.topicPrompts &&
+        prev.topicRes === next.topicRes, // ,
     }
-  }, [topicState.processing, topicState.prompts, topicState.res]);
-
-}
-*/
+  );
+};
